+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@
 require "bundler/setup"
 require "dotenv/load"
 require "beyond_api"
+require "factory_bot"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
@@ -14,6 +15,26 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.include FactoryBot::Syntax::Methods
+
+  config.before(:suite) do
+    FactoryBot.find_definitions
+  end
+
+  config.after(:suite) do
+    session = BeyondApi::Session.new(api_url: ENV["SHOP_URL"])
+    session.token.client_credentials
+
+    products = session.products.all
+    products.embedded.products.each do |product|
+      session.products.delete(product.id)
+    end
+  end
+
+  AppRoot = File.expand_path(File.dirname("ext.rb"))
+
+  load "#{AppRoot}/lib/beyond_api/ext.rb"
 end
 
 unless ENV["CLIENT_ID"].nil? && ENV["CLIENT_SECRET"].nil?